+++ www/helpers.js
@@ -1,25 +1,29 @@
-module.exports = function init(jsUtil, cookieHandler, messages) {
+module.exports = function init(jsUtil, cookieHandler, messages, base64, errorCodes) {
   var validSerializers = ['urlencoded', 'json', 'utf8'];
   var validCertModes = ['default', 'nocheck', 'pinned', 'legacy'];
   var validClientAuthModes = ['none', 'systemstore', 'buffer'];
   var validHttpMethods = ['get', 'put', 'post', 'patch', 'head', 'delete', 'upload', 'download'];
+  var validResponseTypes = ['text', 'json', 'arraybuffer', 'blob'];
 
   var interface = {
     b64EncodeUnicode: b64EncodeUnicode,
-    checkSerializer: checkSerializer,
-    checkSSLCertMode: checkSSLCertMode,
     checkClientAuthMode: checkClientAuthMode,
     checkClientAuthOptions: checkClientAuthOptions,
+    checkDownloadFilePath: checkDownloadFilePath,
+    checkFollowRedirectValue: checkFollowRedirectValue,
     checkForBlacklistedHeaderKey: checkForBlacklistedHeaderKey,
     checkForInvalidHeaderValue: checkForInvalidHeaderValue,
+    checkSerializer: checkSerializer,
+    checkSSLCertMode: checkSSLCertMode,
     checkTimeoutValue: checkTimeoutValue,
-    checkFollowRedirectValue: checkFollowRedirectValue,
-    injectCookieHandler: injectCookieHandler,
-    injectFileEntryHandler: injectFileEntryHandler,
+    checkUploadFileOptions: checkUploadFileOptions,
     getMergedHeaders: getMergedHeaders,
     getProcessedData: getProcessedData,
     handleMissingCallbacks: handleMissingCallbacks,
-    handleMissingOptions: handleMissingOptions
+    handleMissingOptions: handleMissingOptions,
+    injectCookieHandler: injectCookieHandler,
+    injectFileEntryHandler: injectFileEntryHandler,
+    injectRawResponseHandler: injectRawResponseHandler,
   };
 
   // expose all functions for testing purposes
@@ -28,6 +32,7 @@ module.exports = function init(jsUtil, cookieHandler, messages) {
     interface.checkForValidStringValue = checkForValidStringValue;
     interface.checkKeyValuePairObject = checkKeyValuePairObject;
     interface.checkHttpMethod = checkHttpMethod;
+    interface.checkResponseType = checkResponseType;
     interface.checkHeadersObject = checkHeadersObject;
     interface.checkParamsObject = checkParamsObject;
     interface.resolveCookieString = resolveCookieString;
@@ -91,8 +96,26 @@ module.exports = function init(jsUtil, cookieHandler, messages) {
     return obj;
   }
 
+  function checkArray(array, allowedDataTypes, onInvalidValueMessage) {
+    if (jsUtil.getTypeOf(array) !== 'Array') {
+      throw new Error(onInvalidValueMessage);
+    }
+
+    for (var i = 0; i < array.length; ++i) {
+      if (allowedDataTypes.indexOf(jsUtil.getTypeOf(array[i])) === -1) {
+        throw new Error(onInvalidValueMessage);
+      }
+    }
+
+    return array;
+  }
+
   function checkHttpMethod(method) {
     return checkForValidStringValue(validHttpMethods, method, messages.INVALID_HTTP_METHOD);
+  }
+
+  function checkResponseType(type) {
+    return checkForValidStringValue(validResponseTypes, type, messages.INVALID_RESPONSE_TYPE);
   }
 
   function checkSerializer(serializer) {
@@ -165,7 +188,7 @@ module.exports = function init(jsUtil, cookieHandler, messages) {
 
   function checkForInvalidHeaderValue(value) {
     if (jsUtil.getTypeOf(value) !== 'String') {
-      throw new Error(messages.INVALID_HEADERS_VALUE);
+      throw new Error(messages.INVALID_HEADER_VALUE);
     }
 
     return value;
@@ -188,11 +211,44 @@ module.exports = function init(jsUtil, cookieHandler, messages) {
   }
 
   function checkHeadersObject(headers) {
-    return checkKeyValuePairObject(headers, ['String'], messages.INVALID_HEADERS_VALUE);
+    return checkKeyValuePairObject(headers, ['String'], messages.TYPE_MISMATCH_HEADERS);
   }
 
   function checkParamsObject(params) {
-    return checkKeyValuePairObject(params, ['String', 'Array'], messages.INVALID_PARAMS_VALUE);
+    return checkKeyValuePairObject(params, ['String', 'Array'], messages.TYPE_MISMATCH_PARAMS);
+  }
+
+  function checkDownloadFilePath(filePath) {
+    if (!filePath || jsUtil.getTypeOf(filePath) !== 'String') {
+      throw new Error(messages.INVALID_DOWNLOAD_FILE_PATH);
+    }
+
+    return filePath;
+  }
+
+  function checkUploadFileOptions(filePaths, names) {
+    if (jsUtil.getTypeOf(filePaths) === 'String') {
+      filePaths = [filePaths];
+    }
+
+    if (jsUtil.getTypeOf(names) === 'String') {
+      names = [names];
+    }
+
+    var opts = {
+      filePaths: checkArray(filePaths, ['String'], messages.TYPE_MISMATCH_FILE_PATHS),
+      names: checkArray(names, ['String'], messages.TYPE_MISMATCH_NAMES)
+    };
+
+    if (!opts.filePaths.length) {
+      throw new Error(messages.EMPTY_FILE_PATHS);
+    }
+
+    if (!opts.names.length) {
+      throw new Error(messages.EMPTY_NAMES);
+    }
+
+    return opts;
   }
 
   function resolveCookieString(headers) {
@@ -224,6 +280,46 @@ module.exports = function init(jsUtil, cookieHandler, messages) {
     return function (response) {
       cookieHandler.setCookieFromString(url, resolveCookieString(response.headers));
       cb(response);
+    }
+  }
+
+  function injectRawResponseHandler(responseType, success, failure) {
+    return function (response) {
+      var dataType = jsUtil.getTypeOf(response.data);
+
+      // don't need post-processing if it's already binary type (on browser platform)
+      if (dataType === 'ArrayBuffer' || dataType === 'Blob') {
+        return success(response);
+      }
+
+      try {
+        // json
+        if (responseType === validResponseTypes[1]) {
+          response.data = JSON.parse(response.data);
+        }
+
+        // arraybuffer
+        else if (responseType === validResponseTypes[2]) {
+          response.data = base64.toArrayBuffer(response.data);
+        }
+
+        // blob
+        else if (responseType === validResponseTypes[3]) {
+          var buffer = base64.toArrayBuffer(response.data);
+          var type = response.headers['content-type'] || '';
+          var blob = new Blob([ buffer ], { type: type });
+          response.data = blob;
+        }
+
+        success(response);
+      } catch (error) {
+        failure({
+          status: errorCodes.POST_PROCESSING_FAILED,
+          error: messages.POST_PROCESSING_FAILED + ' ' + error.message,
+          url: response.url,
+          headers: response.headers
+        });
+      }
     }
   }
 
@@ -277,7 +373,7 @@ module.exports = function init(jsUtil, cookieHandler, messages) {
     var allowedDataTypes = getAllowedDataTypes(dataSerializer);
 
     if (allowedDataTypes.indexOf(currentDataType) === -1) {
-      throw new Error(messages.DATA_TYPE_MISMATCH + ' ' + allowedDataTypes.join(', '));
+      throw new Error(messages.TYPE_MISMATCH_DATA + ' ' + allowedDataTypes.join(', '));
     }
 
     if (dataSerializer === 'utf8') {
@@ -301,15 +397,16 @@ module.exports = function init(jsUtil, cookieHandler, messages) {
     options = options || {};
 
     return {
-      method: checkHttpMethod(options.method || validHttpMethods[0]),
-      serializer: checkSerializer(options.serializer || globals.serializer),
-      timeout: checkTimeoutValue(options.timeout || globals.timeout),
+      data: jsUtil.getTypeOf(options.data) === 'Undefined' ? null : options.data,
+      filePath: options.filePath,
       followRedirect: checkFollowRedirectValue(options.followRedirect || globals.followRedirect),
       headers: checkHeadersObject(options.headers || {}),
+      method: checkHttpMethod(options.method || validHttpMethods[0]),
+      name: options.name,
       params: checkParamsObject(options.params || {}),
-      data: jsUtil.getTypeOf(options.data) === 'Undefined' ? null : options.data,
-      filePath: options.filePath || '',
-      name: options.name || ''
+      responseType: checkResponseType(options.responseType || validResponseTypes[0]),
+      serializer: checkSerializer(options.serializer || globals.serializer),
+      timeout: checkTimeoutValue(options.timeout || globals.timeout),
     };
   }
 };