+++ www/public-interface.js
@@ -1,5 +1,5 @@
-module.exports = function init(exec, cookieHandler, urlUtil, helpers, globalConfigs) {
-  const publicInterface = {
+module.exports = function init(exec, cookieHandler, urlUtil, helpers, globalConfigs, errorCodes) {
+  var publicInterface = {
     getBasicAuthHeader: getBasicAuthHeader,
     useBasicAuth: useBasicAuth,
     getHeaders: getHeaders,
@@ -28,7 +28,8 @@ module.exports = function init(exec, cookieHandler, urlUtil, helpers, globalConf
     delete: del,
     head: head,
     uploadFile: uploadFile,
-    downloadFile: downloadFile
+    downloadFile: downloadFile,
+    ErrorCode: errorCodes
   };
 
   function getBasicAuthHeader(username, password) {
@@ -143,22 +144,25 @@ module.exports = function init(exec, cookieHandler, urlUtil, helpers, globalConf
     url = urlUtil.appendQueryParamsString(url, urlUtil.serializeQueryParams(options.params, true));
 
     var headers = helpers.getMergedHeaders(url, options.headers, globalConfigs.headers);
-    var onSuccess = helpers.injectCookieHandler(url, success);
+
     var onFail = helpers.injectCookieHandler(url, failure);
+    var onSuccess = helpers.injectCookieHandler(url, helpers.injectRawResponseHandler(options.responseType, success, failure));
 
     switch (options.method) {
       case 'post':
       case 'put':
       case 'patch':
         var data = helpers.getProcessedData(options.data, options.serializer);
-        return exec(onSuccess, onFail, 'CordovaHttpPlugin', options.method, [url, data, options.serializer, headers, options.timeout, options.followRedirect]);
+        return exec(onSuccess, onFail, 'CordovaHttpPlugin', options.method, [url, data, options.serializer, headers, options.timeout, options.followRedirect, options.responseType]);
       case 'upload':
-        return exec(onSuccess, onFail, 'CordovaHttpPlugin', 'uploadFile', [url, headers, options.filePath, options.name, options.timeout, options.followRedirect]);
+        var fileOptions = helpers.checkUploadFileOptions(options.filePath, options.name);
+        return exec(onSuccess, onFail, 'CordovaHttpPlugin', 'uploadFiles', [url, headers, fileOptions.filePaths, fileOptions.names, options.timeout, options.followRedirect, options.responseType]);
       case 'download':
+        var filePath = helpers.checkDownloadFilePath(options.filePath);
         var onDownloadSuccess = helpers.injectCookieHandler(url, helpers.injectFileEntryHandler(success));
-        return exec(onDownloadSuccess, onFail, 'CordovaHttpPlugin', 'downloadFile', [url, headers, options.filePath, options.timeout, options.followRedirect]);
+        return exec(onDownloadSuccess, onFail, 'CordovaHttpPlugin', 'downloadFile', [url, headers, filePath, options.timeout, options.followRedirect]);
       default:
-        return exec(onSuccess, onFail, 'CordovaHttpPlugin', options.method, [url, headers, options.timeout, options.followRedirect]);
+        return exec(onSuccess, onFail, 'CordovaHttpPlugin', options.method, [url, headers, options.timeout, options.followRedirect, options.responseType]);
     }
   }