+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 2.2.0
+
+- Feature #239: add enumeration style object for error codes
+- Feature #253: add support for response type "json"
+- Feature #127: add multiple file upload (thanks SDA SE Open Industry Solutions and nilswitschel)
+
+## 2.1.1
+
+- Fixed #224: response type "arraybuffer" and "blob" not working on browser platform
+
+## 2.1.0
+
+- Feature #216: Support for response type `arraybuffer`
+- Feature #171: Support for response type `blob`
+- Feature #205: Add preference for configuring OKHTTP version (thanks RougeCiel)
+
+## 2.0.11
+
+- Fixed #221: headers not set on Android when request fails due to non-success status code
+
+## 2.0.10
+
+- Fixed #218: headers are used as params on browser platform
+
 ## 2.0.9
 
 - Fixed #204: broken support for cordova-android  < 7.0